+++ alteia/__init__.py
@@ -4,6 +4,6 @@ from alteia.sdk import SDK
 
 __all__ = ('SDK',)
 
-__version__ = '1.0.1'  # must match the version in pyproject.toml
+__version__ = '1.0.2'  # must match the version in pyproject.toml
 
 __path__ = extend_path(__path__, __name__)